Refactor project into src package
